COMPLETE REORGANIZATION
